+++ packages/extension-base/src/services/transaction-service/index.ts
@@ -1474,36 +1474,57 @@ export default class TransactionService {
       this.handleTransactionTimeout(emitter, eventData);
       emitter.emit('send', eventData); // This event is needed after sending transaction with queue
 
+      let isBroadcast = false;
+      let isInBlock = false;
+      let isFinish = false;
+
       rs.send((txState) => {
         // handle events, logs, history
         if (!txState || !txState.status) {
           return;
         }
 
-        if (txState.status.isInBlock) {
-          eventData.eventLogs = txState.events;
-
-          if (!eventData.extrinsicHash || eventData.extrinsicHash === '' || !isHex(eventData.extrinsicHash)) {
+        // Broadcast transaction
+        if (!isBroadcast) {
+          if (txState.status.isBroadcast || txState.status.isInBlock || txState.status.isFinalized) {
             eventData.extrinsicHash = txState.txHash.toHex();
-            eventData.blockHash = txState.status.asInBlock.toHex();
-            emitter.emit('extrinsicHash', eventData);
+            isBroadcast = true;
+
+            if (!isFinish) {
+              emitter.emit('extrinsicHash', eventData);
+            }
           }
         }
 
-        if (txState.status.isFinalized) {
-          eventData.extrinsicHash = txState.txHash.toHex();
-          eventData.eventLogs = txState.events;
-          // TODO: push block hash and block number into eventData
-          txState.events
-            .filter(({ event: { section } }) => section === 'system')
-            .forEach(({ event: { data: [error], method } }): void => {
-              if (method === 'ExtrinsicFailed') {
-                eventData.errors.push(new TransactionError(BasicTxErrorType.SEND_TRANSACTION_FAILED, error.toString()));
-                emitter.emit('error', eventData);
-              } else if (method === 'ExtrinsicSuccess') {
-                emitter.emit('success', eventData);
-              }
-            });
+        // Transaction in block
+        if (!isInBlock) {
+          if (txState.status.isInBlock || txState.status.isFinalized) {
+            eventData.blockHash = txState.status.asInBlock.toHex();
+            eventData.eventLogs = txState.events;
+            isInBlock = true;
+          }
+        }
+
+        // Transaction finished
+        if (!isFinish) {
+          if (txState.status.isInBlock || txState.status.isFinalized) {
+            if (!eventData.extrinsicHash) {
+              eventData.extrinsicHash = txState.txHash.toHex();
+            }
+
+            txState.events
+              .filter(({ event: { section } }) => section === 'system')
+              .forEach(({ event: { data: [error], method } }): void => {
+                if (method === 'ExtrinsicFailed') {
+                  eventData.errors.push(new TransactionError(BasicTxErrorType.SEND_TRANSACTION_FAILED, error.toString()));
+                  emitter.emit('error', eventData);
+                  isFinish = true;
+                } else if (method === 'ExtrinsicSuccess') {
+                  emitter.emit('success', eventData);
+                  isFinish = true;
+                }
+              });
+          }
         }
       }).catch((e: Error) => {
         eventData.errors.push(new TransactionError(BasicTxErrorType.SEND_TRANSACTION_FAILED, e.message));